refactor: Enhance command handling with new throttle and turnout features
