+++ src/turnouts.mjs
@@ -1,37 +1,43 @@
 import log from './utils/logger.mjs'
+import layout from './layout.mjs'
 
-async function handleTurnout(turnout) {
+export async function handleTurnout(turnout) {
   try {
-    const command = await turnoutCommand(turnout)
-    return command
+    const conn = layout.connections()?.[turnout.device]
+    if (!conn?.isConnected) {
+      log.error('Device not connected', turnout.device)
+      return
+    }
+    const command = turnoutCommand(turnout)
+    const layoutDevice = layout
+      .devices()
+      ?.find(({ id }) => id === turnout.device)
+    log.log(
+      'handleTurnoutCommand',
+      turnout,
+      command,
+      conn?.isConnected,
+      layoutDevice
+    )
+    if (layoutDevice?.connection === 'usb') {
+      await conn.send(conn.port, JSON.stringify([command]))
+    } else if (layoutDevice?.connection === 'wifi') {
+      await conn.send(conn.topic, JSON.stringify(command))
+    }
   } catch (err) {
     log.error('[COMMANDS] turnoutCommand', err)
   }
 }
+
 export function turnoutCommand(turnout) {
   try {
     let commands = []
     switch (turnout?.type) {
       case 'kato':
-        commands.push({
-          device: turnout.device,
-          action: 'turnout',
-          payload: {
-            turnout: turnout.turnoutIdx,
-            state: turnout.state,
-          },
-        })
+        commands.push(katoCommand(turnout))
         break
       case 'servo':
-        commands.push({
-          device: turnout.device,
-          action: 'servo',
-          payload: {
-            servo: turnout.turnoutIdx,
-            value: turnout.state ? turnout.straight : turnout.divergent,
-            current: !turnout.state ? turnout.straight : turnout.divergent,
-          },
-        })
+        commands.push(servoCommand(turnout))
         break
       default:
         // no op
@@ -43,7 +49,40 @@ export function turnoutCommand(turnout) {
   }
 }
 
+function katoCommand(turnout) {
+  return {
+    device: turnout.device,
+    action: 'turnout',
+    payload: {
+      turnout: parseInt(turnout.turnoutIdx),
+      state: turnout.state,
+    },
+  }
+}
+
+function servoCommand(turnout) {
+  return {
+    device: turnout.device,
+    action: 'servo',
+    payload: {
+      servo: parseInt(turnout.turnoutIdx),
+      value: parseInt(turnout.state ? turnout.straight : turnout.divergent),
+      current: parseInt(!turnout.state ? turnout.straight : turnout.divergent),
+    },
+  }
+}
+
+export async function handleTurnoutChange(snapshot) {
+  snapshot.docChanges().forEach(async (change) => {
+    if (change.type === 'modified') {
+      console.log('handleTurnoutChange', change.type, change.doc.data())
+      await handleTurnout(change.doc.data())
+    }
+  })
+}
+
 export default {
   turnoutCommand,
   handleTurnout,
+  handleTurnoutChange,
 }
